The CDLP variant that does not require a sorting step is now the default. The previous CDLP default variant is now called LAGraph_CDLP_withsort.
